Move users_at_location lookup to form
